+++ back/src/main/java/com/udteam/miristock/entity/CommentEntity.java
@@ -17,7 +17,7 @@ public class CommentEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long commentNo;
+    private Integer commentNo;
 
     private String commentContent;
